requirements.txt
scratchattach
python-dotenv
